Dharma-Platform-Live-Dashboard-master/services/data-collection-service/requirements.txt
# Data Collection Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data collection libraries
tweepy==4.14.0
google-api-python-client==2.108.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
telethon==1.32.1

# Database and caching
pymongo==4.6.0
redis==5.0.1

# Message queue
kafka-python==2.0.2

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
httpx==0.25.2